docs: move file, rename folder
