+++ src/App.tsx
@@ -2,17 +2,17 @@ import { Route, Routes, useLocation } from "react-router-dom";
 import OnboardingConcernPage from "./services/onboarding/additional/concern";
 import OnboardingEmotionPage from "./services/onboarding/additional/emotion";
 import OnboardingCompletePage from "./services/onboarding/complete";
-import HomePage from "./services/home";
+import HomePage from "./services/home/chat/index.tsx";
 import NotFoundPage from "./services/error/404";
 import OnboardingBasicPage from "./services/onboarding/basic";
 import LoginPage from "./services/auth/login";
 import OAuthCallback from "./services/auth/oauth";
 import Layout from "./layout.tsx";
 import { AnimatePresence } from "framer-motion";
-import PhqPage from "./services/home/test/phq/index.tsx";
+import PhqPage from "./services/home/test/PHQ/index.tsx";
 import SettingPage from "./services/setting/index.tsx";
-import GadPage from "./services/home/test/gad/index.tsx";
-import PssPage from "./services/home/test/pss/index.tsx";
+import GadPage from "./services/home/test/GAD/index.tsx";
+import PssPage from "./services/home/test/PSS/index.tsx";
 import RootRedirect from "./services/redirect/rootRedirect.tsx";
 import { useTokenMonitor } from "./hooks/auth/useTokenMonitor";
 import ReportMainPage from "./services/report/main/index.tsx";
