+++ src/services/home/test/SolutionCard/solutionLevelMap.ts
